Add support for custom ACS URL in CreateAuthnRequest() and ParseResponse()
The URL can now be customized using `WithAssertionConsumerServiceURL()`
in both functions.

To validate the behavior I added a short test for `ServiceProvider.ParseResponse`.
It only checks the error to make sure `WithAssertionConsumerServiceURL()`
for now but can be extended in the future.

Also fix a docstring and gives the custom clock from `WithClock()` to the
internal parser.