Remove NameIDFormat default
NameIDFormat is

	Zero or more elements of type anyURI that enumerate the name identifier
	formats supported by this system entity acting in this role. See Section
	8.3 of [SAMLCore] for some possible values for this element.

Vault (and I think Boundary also) do not need it to be an email, the name
is used as an opaque string.

I suppose most public IdP will have an email for their user but private
ones may not have one and use an employee ID or a username instead.

I think it's best to keep the default configuration empty to be compatible
with such systems.