OIDC private key JWT / client assertion (#155)
* add oidc/clientassertion package

for signing JWTs to send as client_assertions
ref: https://oauth.net/private-key-jwt/

* add WithClientAssertionJWT Option

to send a signed JWT as request's client_assertion
ref: https://oauth.net/private-key-jwt/

* test WithClientAssertionJWT

* add changelog, fix test

* s/ClientAssertion/JWT/g in clientassertion pkg

* move options to options.go

* go-jose/v3->v4

v4 is more sensitive to HMAC length

* more and different option validation

* WithClientAssertionJWT accepts a Serializer

instead of a string, so it's harder to use incorrectly.
JWT implements Serializer, but folks may provider their own.

* protect kid from headers

* add ExampleJWT test

* brief package docstring, rename a const

* add copyright headers

* missed a rename

* test the new method on test provider

* godocs and ops everywhere!

* privatize a couple things, const KeyIDHeader

* errors go in error.go

* error if missing kid

* fix obvious mistake

it is getting late in the day!

* fix less obvious but still pretty obvious error

if i'd just run my own tests (:

* add a couple missing err ops

* pr feedback bits

* clarify hmac min len

* refactor to NewJWTWithRSAKey and NewJWTWithHMAC

* remove most single-use private methods

i was a bit overzealous with my tiny methods.

i couldn't quite bring myself to delete signer() :P

* more lovely PR feedback